backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pandas==2.0.3
requests==2.31.0
langchain>=0.0.335
langchain-core>=0.1.17
langchain-community>=0.0.13
langchain-huggingface>=0.0.2
langchain-openai>=0.0.5
faiss-cpu==1.7.4
python-dotenv==1.0.0
huggingface-hub>=0.30.0
sentence-transformers>=2.2.2
openai>=1.3.0
tiktoken>=0.5.1
supabase>=2.15.3
pyjwt>=2.10.1
passlib>=1.7.4
python-jose>=3.3.0
python-multipart>=0.0.6
